# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking System API",
        "version": "1.0"
    },
    "paths": {
        "/api/events": {
            "get": {
                "tags": [
                    "Event API"
                ],
                "summary": "Получить список событий",
                "description": "Получите список всех мероприятий.",
                "operationId": "7058ab7c89d20c752b6afd6d808f6526",
                "responses": {
                    "200": {
                        "description": "Список мероприятий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Show #1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Event API"
                ],
                "summary": "Получите событие для конкретного мероприятия",
                "description": "Получите список событий для конкретного мероприятия по его идентификатору.",
                "operationId": "74dc821f168250ed86869f5cc643ce8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор мероприятия",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список мероприятий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 46
                                                    },
                                                    "showId": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2019-08-22 20:26:38"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Мероприятий не найдено"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/events/{id}/places": {
            "get": {
                "tags": [
                    "Event API"
                ],
                "summary": "Получите места для конкретного мероприятия",
                "description": "Получите доступные места для конкретного мероприятия по его идентификатору.",
                "operationId": "f9eab67b2ce1b40b6e512bde09bcd3ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор мероприятия",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список доступных мест для проведения мероприятия",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "x": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 0
                                                    },
                                                    "y": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 0
                                                    },
                                                    "width": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 20
                                                    },
                                                    "height": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 20
                                                    },
                                                    "is_available": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Мероприятие не найдено или места недоступны"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/events/{id}/reserve": {
            "post": {
                "tags": [
                    "Event API"
                ],
                "summary": "Забронируйте места для мероприятия",
                "description": "Забронируйте места для конкретного мероприятия по его идентификатору.",
                "operationId": "461495b34fd13ad0cdb70a27340cd2cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор мероприятия",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "places": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Подтверждение бронирования",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "object"
                                        },
                                        "response.success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "response.reservation_id": {
                                            "type": "string",
                                            "example": "5d519fe58e3cf"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос на бронирование"
                    },
                    "404": {
                        "description": "Событие не найдено"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Event API",
            "description": "Операции, связанные с мероприятиями"
        }
    ]
}